requirements.txt
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.3
requests==2.31.0
shopify==5.5.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
mypy==1.7.1
black==23.11.0
isort==5.12.0